requirements.txt
asgiref
Django
djangorestframework
psycopg2-binary
pillow
python-dotenv
django_filter
djangorestframework-simplejwt
drf-yasg
stripe





